[ refactor ] 서비스계층 추상화 ( #66 )
